refactor: split faucet components
