+++ packages/app/src/styles/index.css
@@ -11,10 +11,11 @@
 @import "./components/faucet-widget.css";
 @import "./components/header.css";
 @import "./components/input.css";
-@import "./components/layout.css";
+@import "./components/main-layout.css";
 @import "./components/popover.css";
 @import "./components/tooltip.css";
 @import "./components/toast.css";
+@import "./components/welcome-page.css";
 
 @import "tailwindcss/utilities";
 @import "./utilities.css";
